backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
pydantic==2.10.*
httpx==0.28.*
python-dotenv==1.0.*
openpyxl==3.1.*
python-multipart==0.0.20
sqlalchemy[postgresql]>=2.0
pgvector>=0.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
